ModInfo crc reworking and fixes - moved to the data model - WIP:
setmtime was doing a very good job of not recalculating the
crc - only it did its job too well, missing updates.

This commit at least makes sure that updates (like esm flip see below)
_inside_ Bash will correctly update the crc, by adding a crc_changed
param in setmtime. I also added dedicated methods using the cached crc.
This resulted in dropping the system call in cachedCrc - we had O(n)
of them to the number of mods on each RefreshUI. The deal is that by the
time we refresh the UI (or even more sort the displayed items) the data
model is already refreshed, so we can use the cached value directly
here. Still the situation with the FileInfo(s) APIs is suboptimal -
I should be able to control when an update is needed to the crc in
needs_update() but that proved tricky. I was not really able to spell
the parameters to _reset_cache, hence the "and hopefully updated". So
_reset_cache would ideally recalculate the crc but it's also called in
init via refreshFile (rename !) > factory so it would result in initial
ModInfos refresh recalculate all the crcs. On the other hand,
if for instance size and mtime stay the same while crc changed this
won't be caught here - one such case would be esm flip, done outside
Bash (with no change in modification time). Note that ctime will be
changed in this case (at least on windows) so the file _will_ be marked
as updated. WIP, see #336 , #377