Path - lazily compute seldom used caches - comments:
I took the trick from pathlib, for instance:
https://github.com/python/cpython/blob/7a113a0cbf545588d61286fcc0e89141cf211735/Lib/pathlib.py#L725-L732
Still maybe caching those is no use for speed, see:
http://stackoverflow.com/questions/1479979/case-insensitive-comparison-of-sets-in-python/1480230#comment1332537_1480230
Note in particular this commit does not reduce memory usage, probably
due to python using the same string for _cext as for _ext (as extensions
are lowercase anyway in my data/ dir) ? The cext property is _not_ used
in the memory intensive path of BAIN initial refresh (only used in
installers and converters dir refreshes but that's a few 100 files), so
_cext cahce is not populated for the data_SizeCrc_Date keys for instance
(which are several 10ks in my case).

Under #368.