Reworking setOblivionVersion to drop modInfos refresh:
This was triggered by me realizing that I needed to refresh with
load_order.Unlock, but decided to solve it at the root. We should
eliminate _all_ occurrences of refresh outside the data model (under

- ghosting. Ghost state won't be preserved for renamed master if
autoGhost is False, but that's not important. What's important is
(further) reworking lo API to always unghost if needed (eliminate
setGhost calls).
- our pre alpha file system APIs. That's partly due to Path API (#368)
and partly to the hacked in SHFileOperation that should become the norm,
but error handling is a huge issue (see the ugly _retry in this context)
Should be moved inside env - #241

On the bright side: self.factory dropped, performance increased by an
order of magnitude and basher.saves_links.Saves_Profiles._ProfileLink
swapPluginsAndMasterVersion > setOblivionVersion call should now
properly refresh (oops).