Get default icon for exe files if exe bundles no icon:
I initially just switched to ExtrcaticonEx - just query if there are
embedded icons, instead of loading the icon just to destroy it:

@@ -224,14 +224,7 @@ def init_app_links(apps_dir, badIcons, iconList):
             icon, idex = icon.split(u',')
-            if icon == u'' and win32gui is not None:
-                if target.cext == u'.exe':
-                    # Use the icon embedded in the exe
-                    # noinspection PyUnresolvedReferences
-                    try:
-                        hIcon = win32gui.ExtractIcon(0, target.s, 0)
-                        win32gui.DestroyIcon(hIcon)
+            if icon == u'':
+                if win32gui is not None and target.cext == u'.exe' and \
+                    win32gui.ExtractIconEx(target.s, -1):
+                        # -1 queries num of icons embedded in the exe
                         icon = target
-                    except pywintypes.error: # complains error is unresolved
-                        #(1813, 'ExtractIcon', 'The specified resource type
-                        # cannot be found in the image file.')
-                        icon = u'' # Icon will be set to a red x further down.
                 else:

This got me rid of a lot of ugly code but I would still get a red x for
exes that bundled no icon. As seen I let it go through
`else: icon, idex = _get_default_app_icon(idex, target)`
to get the generic exe icon, but this failed miserably, as filedata
variable was set to u'%1' and this couldn't be split in
`icon, idex = filedata.split(u',')`

After a lot of fighting with SHGetFileInfo horrid API, I went ahead and
hardcoded the ShellDll generic application icon.