Removing messages handling from _fix methods in games.py:
It's not the business of this layer to print messages - however the
fix calculations should only be performed once (so we could not have
load_order layer recalculate added, removed etc). This lead to this WIP
design that however already simplifies the guts of games.py (much
welcome) - WIP in the sense that FixInfo is not the pinnacle of
elegance, messages need rewrite and we badly need tests.
Note I moved check for duplicates before checking for more than 255
plugins in _fix_active_plugins.

Under #309.