Common code in Installer#install - WIP:
Can you spot the bug in InstallerProject#install ?

We used to timestamp the stageFull - this code used to be much more
complicated. We have to do away with timestamps altogether (done later
in this branch).
@@ -1408,3 +1408,3 @@ def install(self, destFiles, progress=None):
                 srcFull.copyTo(stageFull)
-                if bass.reModExt.search(srcFull.s): timestamps(srcFull)
+                if bass.reModExt.search(srcFull.s): timestamps(stageFull)
                 data_sizeCrcDate_update[dest] = (size,crc,stageFull.mtime)
Renamed unpackDirJoin to srcDirJoin to have the install() overrides
easily compare.
Took the opportunity to streamline and comment code in InstallerData
_install(), proved rather tricky due to progress bars needing careful
updates of index variable.