[!]Fixup - using ObCollection would change the current working dir:
Happily we used it in a context manager so it was easily monkey patched,
but CBash must be fixed - https://github.com/wrye-bash/CBash/issues/20
Idea from: http://stackoverflow.com/a/13197763/281545
The change of cwd would have unpredictable consequences - probably the
eternal Installer._move bug, that was only affecting oblivion but not
Skyrim (no CBash), was due to this. However it was maybe masking bugs
also, so things may break - hence the !
ObCollection was using __slots__ while not a new style class, fixed.
import struct was unused

Signed-off-by: MrD <the.ubik@gmail.com>