Purge all hardcoded backslash path separators
They should be replaced by either calls to Bash's own path system,
calls to os.path.join, or at least os.sep to make them (more) os agnostic.

The code isn't super pretty because I mostly went with a one-to-one
simple replace instead of refactoring it more thoroughly. This is merely
a basic cleanup to make Bash more runnable on Linux while hopefully not
blowing up on Windows in the process. Needless to say, this isn't ready
for prime time quite yet.

Still untouched are stuff in the game/constants.py files and some things
either Windows-specific or where I suspected it might be some kind of
internal paths instead of proper file system paths.

Utumno: I removed the handling of localSave as this should be written
to the game ini as a windows path (see commented out code in bosh).

This needs a lot of thought in general due to paths saved in settings as
windows paths (see BAIN especially) and comparisons performed between
paths. Probably normpath internally?

Under # 243

# Conflicts:
#	Mopy/bash/bosh/bain.py
#	Mopy/bash/game/__init__.py