Initial attempt at sizers API
Infernio: These are backwards-compatible portions of one of the
de-wx'ing commits by nycz, working towards wx3 compatibility in this
case. Squashed notes follow:

Proof of concept to replace sizers

Inspired by Qt but designed with the intention to fit well with Bash's
current usage of sizers. Every current use of sizers would obviously
need to be rewritten but the goal would be for the resulting code to be
much smaller and/or easier to understand.

Also very very Very much still just a proof of concept. I've only tried
this in heavily refactored code in frames.py.

WIP Wrap and move event binding to balt

A proof of concept that isn't particularly pretty but works well enough.
On the other hand, if all wx objects will one day be entirely abstracted
away, this might need to be made entirely different, or made as a part
of some base GUI object.

Either way, making events and event binding not depend on wx will be
tricky.

Infernio: Left out frames.py, since the iewin stuff is missing.