Random refactoring in wizards:
- I had to use LowerDicts here, chopping off unneeded methods
- PackageList was unused
- Note pruned except: pass

Mopy/bash/basher/constants.py:
Mopy/bash/basher/installer_links.py:

As reported in the forums on a clean start Bash would complain that:

installer_links.pyo  334 _get_size_and_pos: Saved Wizard position \
((-1, -1)) was not a tuple (<class 'wx._core.Point'>), reverting to \
default position.

This nonsense was due to the fact that the default value was not a
tuple - fixed, but still all this code needs revisiting.