Refactor RecordHeader code:
Monkey patch to handle form versions that became important with SSE.
Monkey patch cause RecordHeader features the anti-pattern of defining
different attributes in its init based on record type (YAK!). The right
way would of course be a record header factory and subclasses, but this
needs a complete overhaul of the records code, which I haven't got round
to refactoring - and it's a full release cycle work, be my guests.
It's a monkey patch in another sense, namely we need to refactor game
handling (see #358) - it does make sense to have different Header
classes per game but as it is I went for eliminating this, and moving
processing in game.init. That's the most efficient way (without
rewriting everything). Once we have a GameInfo class we should revisit
this. The `esp` class should be eliminated - although it seems more
natural to have topTypes and co in there, they are however used in
RecordHeader and moved there for locality of reference.

We need to add more comments - note that in GRUP fromVersion never seems
to be set.

Mopy/bash/game/__init__.py:
The esp.header class attribute was unused - although defining the slots
in RecordHeader and overwriting this for oblivion would be a way to
handle game differences.

Closes #342