Remove Mopy/bash/bish.py:
It has fallen out of sync with main development since forever, and it's
not related to the rest of the code. Keeps coming up in refactoring
results. Feel free to take up and fix in another repo under Wrye Bash.
You can always take it out of the history

$ git show 83f1f685d6e5619b65e0921b7eb21b1da49b1693:Mopy/bash/bish.py > bish.py

Under #173