Bye _refreshInfoLists:
This opens the way to new_info API. Note we used to recalculate info
lists on every addition of a modInfo - so for all mods on first refresh
This code is very difficult to follow and has surely bugs in edge cases,
but as far as I can see the imported and merged were used always after
the ModInfos refresh was complete, so it should be ok. To see usages,
cf parent commit.

Under #336