Homogenize AFile overrides init params:
Namely pass the full path in, as opposed to dir and filename. This:

- corrects some ugly syntax warnings in core implementation
- gets us rid of the IniInfo#new_info override

With this, AFile reaches a proud beta

Note client code does not have to split dir and filename anymore.

ini_info_factory, load_cache='Ignored' is to get rid of the IniInfo
override of new_info - in this (edge) case I allow myself a suppress
warning.
Finalized in next commit - abs_path should be used exclusively in
attribute accesses.

Under #336