Copy over formVersion from record on dumping it:
Instead of always setting it to latest. `if extra1 == 0` should
only be the case for the TES4 record header. This was a showstopper
for 307.beta2 (see #403), and as discussed extensively in #342 we
should actually exclude those records all together from the patch. As
this is far from trivial we at least fix the BP freezing the game.

Closes #342
Closes #403