WIP - game check, decoupling barb from bash and partly from bosh:
The code in bash.py for error checks will blow - bass.settings is still
None. Fixed later.
This is WIP so don't pay much attention to the details - still probably
one of the most important of the commits on this branch - note:

- dropping imports in bass
- dropping ini handling from bass
- drop super ugly/buggy reinitialization of data structures in barb, at
the expense of adding reinit to bush. Initialization will be further
worked in next commits - placing restore settings correctly in boot
sequence is one of the goals of this merge.