WIP - decouple barb from bush and bosh:
I moved getLocalSaveDirs into initialization - this decouples barb from
bosh finally - plus we have most of dirs initialization in the same
place (moved init mopy dirs there too, along with the cd logic).

A note for aspiring architects - introducing fsName parameter is not
actually just a trick for decoupling the core modules - it's good
practice, as it makes clear what info we need to restore/backup. Plus
internal state should be kept to a minimal - the Java immutable class
paradigm.

Note that new_bash_version_prompt_backup and backup_filename depend on
bass.settings - we need a new BashBootInfo.dat to fix this and to keep
the language info - that's under #26