Monkey patch for undecodable inis:
See:
https://afkmods.iguanadons.net/index.php?/topic/4966-wrye-bash-all-games/&do=findComment&comment=173759

__init__.py 1608 _refresh_ini_infos: Failed to read C:\GAMES\TESIV\Oblivion\Data\INI Tweaks\TamagoClub, ~Old Settings [TamagoClub115c_Rev4_Alpha1.7z].ini
Traceback (most recent call last):
  File "bash\bosh\__init__.py", line 1605, in _refresh_ini_infos
    oldInfo = self.factory(tweak_path)
  File "bash\bosh\__init__.py", line 1491, in ini_info_factory
    INICount = IniFile.formatMatch(fullpath)
  File "bash\bosh\ini_files.py", line 63, in formatMatch
    stripped = cls.reComment.sub(u'',line).strip()
UnicodeDecodeError: 'ascii' codec can't decode byte 0xff in position 0: ordinal not in range(128)

That is bigtime monkey patch as we need to add a corrupted data
structure to INIInfos too plus we need to attempt decoding it - well
we can't allow a startup crash on the other hand.
Renamed _refresh_infos to _refresh_ini_tweaks for readability.