[SSE] Account for truncated records in Skyrim.esm
The Bethesda master files ship with some records that have form versions
< 44. Bash needs to be able to parse these properly, since converting
the official masters is not possible.

To do this, we unpack them using the old format, set as many fields as
possible using those vales, and then fill out any missing fields using
the default values. This obviously requires that the default values
are set, which this commit also does (see the MelWeapCrdt change at
the bottom).