WIP: Add fomod installer support:
Adds some initial FOMOD support to BAIN. Note the WIP: in the commit
msg does mean WIP! This does not integrate well with all of BAIN's
features yet (especially noticable problems are the Plugin Filter,
skips and strings renaming, all of which plain don't work because
almost all of refreshDataSizeCrc gets short-circuited). We are placing
this on dev anyways because it keeps dragging along and getting in the
way when working on something on another branch - you wouldn't believe
how often this code broke during the course of #190.

Biggest ugly wart is of course the special 'fomod' subpackage. *This
must go.* It's a hack, nothing more than that - work done in the
inf-190-interlude merge thankfully gave us the tools to be rid of this,
but renames and the need to come up with an actual API for switching a
package to 'FOMOD mode' (thereby encapsulating all the difficulties
we've had with random parts of the code not accounting for the
subpackage) still stand in the way.

Option types are not validated - contradicting group and
option types should be undefined behaviour.
However, to simplify, option type always wins the conflict
and we deliver a warning if needed.

As an example, in a SelectAll group a NotUsable option
would not be selected.

Ignored tags:
  - moduleImage
  - fommDependency

Ignored attributes:
  - position, colour - [moduleName]
  - alwaysInstall, installIfUsable - [file, folder]

Utumno: Added docs, added caching for images, some de-wx'ing and renames

Infernio:
Rewrote the entire FOMOD GUI to keep it up to date with and take
advantage of #190, several bug fixes, feature additions and renames

Offer links to mass (de)select options

Offer a table-based view of the final files

IMO the text-based view of the files we're going to install has always
been really hard to parse - plus now that we no longer have a fixed font
it's gotten even worse. Still, kept the text-based view around (it's
like 10 lines of code) and instead added a checkbox to switch to this
neat new view.

Under #380

Co-authored-by: Infernio <infernio@icloud.com>
Co-authored-by: MrD <the.ubik@gmail.com>