Revise active mods limits for esls - hard limit of 4096 for now:
From @Infernio at acc80eb9214e2e18206282e323829f3e4fb9dd0a

So, the current knowledge on this is:

4096 is the theoretical limit on ESLs - that's how much space there
is in the FormID for them (8^4 = 4096).
However, testing shows that the vanilla game gives up at a much lower
limit. That limit can be raised to around 1800 ESLs with Engine
Fixes. At that point, the game runs out of stdio handles to use for
plugin files.
As for how Wrye Bash should handle this - not sure yet. I don't know
the exact number, but my suspicion is 2048 stdio handles - 254
regular plugins = 1794 ESLs, but I'm not sure if anyone has done
proper testing with that. For now, it's probably alright if WB just
sticks to the theoretical limit - but we should keep this in mind
once more testing is done.

For more information, see:

https://afkmods.iguanadons.net/index.php?/topic/5079-plugin-files-and-you-esmeslesp/
https://docs.microsoft.com/en-us/cpp/c-runtime-library/reference/setmaxstdio?view=vs-2017