Hack for inconsistent bsa/b2a path separators:
See discussion in:
https://afkmods.iguanadons.net/index.php?/topic/4966-wrye-bash-all-games/&do=findComment&comment=170410
https://afkmods.iguanadons.net/index.php?/topic/4966-wrye-bash-all-games/&do=findComment&comment=170367

Hack as it's performance intensive, but introducing heuristics won't
work as some ba2 files have _both_ filename separators. Gone down that
path as Path has backslashes normalized so in has_assets we only need
to lower. Would it make sense to only expect backslashes (so lower())
for bsas and normcase() for ba2s? Maybe but it should not make a huge
difference in performance probably so KISS.

Note:

@@ -372,3 +372,3 @@ def _load_bsa_light(self): raise NotImplementedError
     def has_assets(self, asset_paths):
-        return set((u'%s' % a).lower() for a in asset_paths) & self.assets
+        return set(a.cs for a in asset_paths) & self.assets

asset_paths should all be Paths so better use cached cs instead of
calling lower()