Rename ItemLink.help to _help - add menu_help property:
The goal here is to replace unnecessary overrides of Link._initData
with overrides of a new property, ItemLink.menu_help. To that end,
this commit and the one after it renames ItemLink.help to _help to
bring it in line with the other internal ItemLink variable, _text.

This commit runs PyCharm's rename refactoring as a starting point
for manual renaming of everything it missed.

Rename ItemLink.help to _help, Part 2

See the previous commit for Part 1 and the explanation for this
change.

This commit fixes all (hopefully) remaining usages of the old
ItemLink.help in Wrye Bash that were missed in the automatic
refactoring in Part 1.

Removed MenuLink._help

MenuLink doesn't implement ItemLink, so that assignment did
absolutely nothing (and even had a value of 'UNUSED')

Utumno: I squashed the commits by @Infernio (Thanks!) cause if they
were kept separate they would break the menus. Note I wanted to rename
help to _help because (like text) is a very bad name for an attribute,
too generic, used in build in params (like argparse) it _is_ a builtin
and whatnot. Ideal name would be _menu_help but that would cause quite
a few lines to pass the 79 chars limit. Before this commit there were
0 occurrences of _help in the codebase and now `_help` must ONLY be used
for the menus attribute. I edited this commit to wrap a couple lines
that passed the 79 chars limit.

Having a property will help us add more elaborate help messages (for
instance automatically adding the shortcut) and simplifies further the
menus.

Oh here are the next two commits that needed be squashed:

Created ItemLink.menu_help

Also refactored ItemLink.AppendToMenu to use it.

This property will allow us to replace several custom help
definitions in _initData's with overrides of menu_help - see the
next commit for that.

Changed _ProfileLink.help to menu_help:

This class and its subclasses were using their own property called
'help', which overrode the "help" class variable.

Utumno: Pycharm's automated rename missed instance attributes and
properties overrides, we should report it to them (or have I already?)

Under #174, #163, the eternal