Add help texts to all Link subclasses:
This commit adds help texts to the following classes:
 - _LoListLink (this one really needs a better name)
 - _LabelLink
 - Installer_Espm_SelectAll
 - Installer_Espm_DeselectAll
 - Installer_Espm_Rename
 - Installer_Espm_Reset
 - Installer_Espm_ResetAll
 - Installer_Espm_List
 - Installer_Install
 - Installer_ListStructure
 - Installer_Move
 - Installer_OpenNexus
 - Installer_OpenSearch
 - Installer_OpenTESA
 - Installer_Subs_SelectAll
 - Installer_Subs_DeselectAll
 - Installer_Subs_ToggleSelection
 - Installer_Subs_ListSubPackages
 - InstallerConverter_Apply
 - InstallerConverter_ApplyEmbedded
 - InstallerConverter_Create
 - Mod_FullLoad
 - Mod_CreateBOSSReport
 - Mod_AddMaster
 - Mod_CopyToEsmp
 - Mod_FlipSelf
 - Mod_Face_Import

Fixed the label help text

It was incorrectly referring to both groups and ratings as
'ratings'. Changed it to 'labels', which is already used for
the clear command and a few others.

Added help texts for the remaining subclasses:

These were missed in the earlier commits.
This should now cover all Link subclasses. Hopefully.

This commit adds help texts for:
 - _Installers_Skip and all its subclasses
 - Installers_AutoRefreshProjects
 - Installers_RemoveEmptyDirs
 - Installers_RenameStrings
 - Mods_AutoGhost

Was a bit lazy with _Installers_Skip, this results in some semi-ugly
help tooltips for those skip Links. I don't feel like fixing that is
worth the extra lines of code though.