Fix Missing Exterior Sub Block Cells
Problem:
Exterior Cell Blocks dropped while building cell data.

When a Block is added it's allocated an exact space in memory.
This includes its header, as well as any of its sub-blocks.
The memory allocated is used to keep looping.
The function insTell() is used to check if it should keep looping.
insTell() returns True if the current read position is at EOF.
If you're at the last block, then that is the end of the file.
So the last block will always be dropped if you call insTell().
Because it breaks from the cell traverse loop.