Formatting & Docstring Amendments
This commit is a squashed version of the two commits containing
formatting changes to the now refactored BsaFile code.

I kept these as a separate commit since not having them doesn't break
dev, and also reduces the number of changes to look at in the big
refactoring commit.

---------------------------------------------------------------------

reset_oblivion_mtimes: Formatting & Docstring

This commit amends the previous one to include all relevant formatting
changes and a docstring with some information on why the method
exists and what it does.

Specific changes made:
 - Formatted to PEP 8 style.
 - Swapped keys and values in the bsa_times dict.
 - Reorded the individual bsa_times entries to match the order in
   which the BSAs will load (earliest to latest).

---------------------------------------------------------------------

calculate_hash: Formatting & Docstring

Renamed a local variable and the parameter to match as well.
Removed all but the first comment - the other three were incredibly
useless.