Different color for ESLMs - Change NoMerge color to dark yellow:
Mark ESLs with an ESM flag purple (pink for ESL + blue for ESM).

Note that this color is somewhat similar to the one for the NoMerge
tag. This shouldn't really be a problem, since NoMerge tags are not
supported on games that support ESLs, but the next commit will change
the color of NoMerge plugins to avoid the problem entirely.

Change NoMerge color to dark yellow

To avoid a 'conflict' with the new ESLM color, this changes the color
for plugins with the NoMerge tag to a dark yellow color (dark green
from NoMerge + orange / yellow as a warning).

Not strictly necessary since NoMerge is not supported in ESL games, but
it may help avoid some confusion between the two things for people who
play both types of games.

Prioritize BP/Mergeable color over ESM/ESL color

Otherwise .esp files with an ESM flag that could be ESL-flagged will
only appear as masters.

Color BP and mergeable plugins in master lists

This adds hightlighting for them in the saves tab and every plugin's
master list. Fairly minor, just adds some consistency.

Utumno: we should refactor common code in masterlists UI