[TES5, FO4] Fixed leveled list patcher discarding 0s
To see this in action, try building a BP in Skyrim (wihout this commit)
with the 'Scarcity' mod installed. It changes a bunch of LVLD entries
to be 0. Unfortunately, the patcher checked to see if a value was
present via an implicit bool() conversion - and bool(0) == False.
Therefore, none of those LVLD changes got carried over, since WB
thought Scarcity just didn't have any values present there.

This changes MreLeveledListBase.mergeWith to use an explicit comparison
with None instead to determine if a value is present. Should only affect
Skyrim (LE & SE) and Fallout 4, since the other games don't use
MreLeveledListBase.