Remove last Wrye Morph remants
The 'Morph Factions' patcher was only useful for Wrye Morph, but Wrye
Morph support has been dropped for almost two years now - so let's drop
this thing as well.

Also gets rid of the 'ScriptContents' tag for Oblivion, which was not
actually implemented (support for it seems to have been dropped in a
much earlier version of Wrye Bash) and is only described by the readme
as 'Should not be used. Can cause serious issues'.