Disable 'Clean Data Directory' per game
Prerequisite for Enderal, otherwise it crashes on launch after using WB
with it.

Background: Enderal ships an empty folder called 'JCData', which
JContainers requires in order to function - if it's removed, JContainers
will crash the entire game. Normally, this folder contains some dummy
files and example Lua scripts, which is why it never gets removed for
standard JContainers installations.

But Enderal, as already mentioned, ships it empty - JContainers is
fine with this, but WB's 'Clean Data Directory' feature, which is
enabled by default, will remove the folder, causing Enderal to crash.
Also renamed it to 'Remove Empty Directories', since that's a much
clearer name.

I'm not entirely happy with this, but it's the best solution I can think
of.

Under #433