Improve and expand texture hash exclusions
Squashed commit consisting of these 2 commits:

1. Drop MelModelHash in favor of MelNull

MelNull already does what we want - the definitions are almost
identical! The only real difference is that MelModelHash also saves the
texture hashes as an attribute of the record. This is obviously totally
unnecessary - we don't want to do anything with those hashes.

To that end, let's just drop those classes and use MelNull.

2. [SSE] Drop PROJ/NAM2 and BPDT/NAM5

These are texture hashes as well! zilav pointed this out [1], but it
looks like this got missed when the texture hash changes were originally
made. I don't know if WB actually patches or writes these, but better
safe than sorry.

[1] https://github.com/wrye-bash/wrye-bash/issues/403#issuecomment-354839885.