Use FileInfo.get_masters() where possible
Note: This relies on the cosaves rework, specifically on
822c0bd16e5788d1811449810f34edfee16d77a1.

Much easier to find the definition and search usages for than the
previous '.header.masters' code, which is now only used where necessary.
Will also allow us to be much more flexible in the future if masters
become more complex than that.

This has sat on nightly for 5+ months now without a single report, so I'm
fairly confident it's error-free.