Skip BAIN refresh for xEdit caches - better esl qualification message:
For the same reasons we skip refreshes for xEdit backups, we should also
skip xEdit caches. Also edited the README to match.

Turns out that SSE never received the edits to skip the 'SSEEdit
Backups' folder, so that's included as well. Also took this opportunity
to correct the 'SkyProc' directory (which should be 'SkyProc Patchers')
in the documentation.

Mopy/bash/bosh/__init__.py:

Show a reason when ESL-checked plugin is already an ESL

Previously, we just displayed 'ESL Incapable' without any reason below
the file name when 'Check ESL Qualifications' was run on an ESL-flagged
file. This led to confusion for some users who weren't sure if this was
a bug or expected behavior. We now show 'Already ESL-flagged.' as a
reason, which should make this much more obvious.