Rework game selection popup slightly
Fixes issue 441, sort of.
Now offers a scrollbar if too many entries are present. Had to change
the TextCtrl to a StaticText because wx is a pile of *%@! and kept
sizing the TextCtrl incorrectly.

Also took the opportunity to clean this method up a bit (PEP8 variables,
_wx.GROW -> _wx.EXPAND, etc.).

Finally, we now also show the game list sorted alphabetically. While the
order in which the games are found does matter behind the scenes, this
sorting is purely visual and should have no impact (tested with 4
installed games).