Rename all user-facing 'esp/m' strings to 'plugin'
ESP/M is an outdated and inaccurate term now unless you actually want to
exclude .esl files. Since WB deals with .esl just fine now, let's drop
this outdated terminology in favor of 'plugin', which should be much
more future-proof.

This notably affects the 'Esp/m Filter' in the Installers Tab, which is
now the 'Plugin Filter'. Also corrected a random spelling mistake in the
Sub-Packages section of the Installers Tab.

Only touches GUI and readmes (and some random comments), didn't want to
bother with changing any of the variables and classes with 'Espm' in
their names, especially since that carries the risk of behavior changes.