belt: espm -> plugin, PEP8 all affected areas
In preparation for the deprecation of the keywords with 'Espm' in the
name. Should have no function changes, but you never know. This gets the
PEP8 nightmare that is belt at least somewhat under control.

Added some underscores to purely internal methods too, but not to fn_
and kwd_ methods, since those kind of serve as a pseudo-API, denoting
which methods are used to implement interpreter behavior.