Make the doc browser fully functional
Back/Forwards buttons are now properly enabled / disabled only when we
actually have a page to go back / forwards to. This prevents the crash
that was previously happening if you tried to go forwards when there was
nothing to go forwards to.

Wrapping WebView would be nice and would allow us to offload some
complexity, as well as bringing the same improvements to the mod
checker - done in next few commits.