Allow components as parents
Introduces a new method, _AComponent._resolve(), which resolves an
object down to a wx object (returning either the object itself or
_native_widget for components). We then call this method at the wx
boundary when constructing the wx objects, allowing either wx objects
or components as parents.

Can and should be dropped in favor of simply returning _native_widget
for everything once wrapping is completely finished.