Rework game version calculation
The impetus for this was that the version shown for Enderal was 1.9.32,
which is obviously incorrect (that's Skyrim's version, Enderal's is
1.5.7 right now). Turns out that the cause is, once again, that a
variable in GameInfo is doing too many things. In this case, launch_exe
(which is *already* split off from exe!) should really only determine
which file to use when launching the game - and *nothing* else. So, a
new variable had to be introduced (we can't just use game_detect_file,
that wouldn't work for Skyrim - it has SkyrimLauncher.exe as
game_detct_file, but the version of the launcher is 1.3.22, not the
correct 1.9.32).

That was enough to correct the wizards and the FOMOD installer. However,
the game launch button was STILL showing the incorrect version. So,
Game_Button had to be slightly rewritten to take a new argument and use
that to calculate its version. Also cleaned up some minor issues along
the way (e.g. the xSE button always showing the version, even when
versions are disabled via settings - was incidentally an open TODO in
issue #174)

Mopy/bash/env.py:

Fix BOSS registry detection not working on 64bit

This was quite the hassle. Turns out that env.get_registry_path wasn't
actually working correctly on 64 bit systems. If run on such a system,
it would ignore keys in the form of Software\xyz and instead search
for Software\Wow6432Node\xyz and Software\Wow6432Node\Wow6432Node\xyz,
which is obviously not what we want. This broke BOSS detection, since
BOSS' registry key is HKLM\Software\BOSS, no Wow6432Node involved.

Squashed in here because it contained a fixup for the first commit.

Under #358, #174.