Pluck some low-hanging ur'' fruit
Many of these were entirely unneeded since the literals in question
didn't even contain escape sequences. Exceptions are the constants for
FO3 and FNV, where I used vim to escape the backslashes, and
GameInfo.pklfile, which was manually escaped.

Track with this regex, and filter by file mask *.py: ur('|")[^,]

Before this commit: 637 in string constants, 5 in comments
After this commit:  107 in string constants, 0 in comments

Almost all remaining ones are in regexes with plenty of backslashes,
which is very annoying.

Under #460