Port to 64bit
Squashed version of a bunch of commits to keep them from breaking dev.
There were some arcane crashes reported with this initially, but there
haven't been any for months now - seems installing the vcredist fixed
that for good (which means it was user error, of course).

Fix TaskDialogIndirect calls failing on 64bit

It was a packing issue because Windows is stupid.

Fix building for 64bit

Moved all packages to 64bit (only the wheels and loot needed to be
changed).

Changed the manifest to look for the 64bit dlls.

Removed the two 64bit dll exclusions from setup.py

Infernio:
Exclude the two new DLLs that now get included

Not sure if this is needed, but those DLLs now show up in the built EXE,
and one of them seems to be MSVC redist, so...

Fix 'Scan for Dirty Edits' erroring

I think this illustrates why we should keep the static typing, made
identifying the problem pretty simple :)
That said, my C knowledge is terrible, so this might actually be wrong.

Also dropped unused windll import.

Various readme changes, drop Gohlke links

Those are temporary, so we have to host mirrors ourselves. Obviously
temporary, which is why I kept them on a branch that we can later drop.
Edited the readmes to drop the '32bit is needed' notes and outdated
things (e.g. if you get a MemoryError now, something is *severely*
wrong with your system :P).

Update all bundled binaries, switch to 64bit RRR

1. 7zip to 19.00 64bit
Note that I used the 64bit version, since we're switching to 64bit
anyways and Igor has claimed it's ~10% faster:
https://sourceforge.net/p/sevenzip/discussion/45797/thread/e164d46a/

2. LZMA SDK to 19.00
To bring it up to date with 7z above. Still the 32bit version, there
doesn't seem to be a 64bit one.

3. pngcrush to 64bit
Actually seems to be quite a bit faster, otherwise I wouldn't have
bothered.

Closes #481.

Co-authored-by: Infernio <infernio@icloud.com>