Fixups for Enderal
1. A custom pkl file _is_ needed - there are some differences between
   which GMSTs are present in Skyrim's and Enderal's Skyrim.esm files.
   Throw in a set.difference debug print if you want a clear diff.
2. Enderal had a GMST record with an empty eid, which WB was not at all
   equipped to handle. xEdit defaults to int, so we will now do that as
   well.

Fixes the crash when building a BP with default settings. Included
gmstEids to save other people the hassle in the future. Also renamed one
occurence of 'format', since that's a builtin.
Pushing this onto dev quickly, since regression fixes (which I guess
this technically is? At the very least, inf-433-enderal is broken on
dev) should land as close to the problematic commit as possible.

Under #433