Rework CTDA handling
This eliminates four unnecessary constants that had to be copy-pasted
for every game.

Also took the opportunity to PEP8 conditionFunctionData and to sort the
CTDA for Oblivion. Needs changes to the xEdit xSE plugin, so that it can
dump things in this new format, which I've done in
https://github.com/TES5Edit/TES5Edit/pull/727.

Note the similarities between FNV <-> FO3 and Skyrim <-> FO4 - however,
I kept these entirely seperate to make contributing updated definitions
painless.