TES5: Fully decode IMAD
Added some utility classes to brec for this, since those will come in
handy for FO3/FNV as well.

This is a fairly pointles commit funcionality-wise - we don't patch
IMAD, and the previous definition was good enough to load the record
(for ESL / merge checking). However, the real win here is getting rid of
14 usages of dumpExtra, a really ugly hack in the record loading code.