TES5: Implement EPFD for PERK properly
Still not 100% perfect - see the TODOs - but it should read it properly
for most cases now. Just like the rest of this branch, completely
untested.

Also features a refactoring of MelString.dumpData and
MelUnicode.dumpData into ModWriter.write_string, because I needed that
code and wasn't going to copy-paste it again.