TES5: Properly implement DOBJ
Now properly reads the < 8 bytes of noise at the end and dumps them out
when writing. xEdit will do the same, unless the record has been
modified, in which case it will just cut the extraneous bytes off.
Didn't do that here, shouldn't hurt to keep them just in case.

This fixes the Export > Editor Ids... link, which was previously
erroring when it was used on Skyrim's Update.esm or Dawnguard.esm.