Improve 'Test Full Record Definitions'
This link is invaluable for figuring out which parts of the record
definitions are still broken, but the way it provides feedback is far
from ideal. This makes it at least show a popup indicating success or
failure.

Had this sitting around for a long time on a half-forgotten branch.

Utumno:
Some label renames

Leaked from future 480-pt2 merge, under #480

Co-authored-by: MrD <the.ubik@gmail.com>