FO3, FNV: Patch COBJ, clean up some constants
COBJ is unused, but we also patch INGR and APPA for Skyrim, so why not
patch this too? Straightforward record, can be hit by 3 different
patchers.

Also threw in some cleanups in preparation for the next commit. Note
that I dropped SCRI and TNAM from the LVL* definitions - they were
leftovers from Oblivion.