FO3, FNV, TES5: Update Import Actors
This one was pretty tough, due to how much it patches. For FO3 and FNV,
it should patch exactly what valda's version patched (with one
difference, see paragraph below).
Diff is that big because of the readmes, especially Actors.ACBS and
Actors.Stats needed a lot of documentation.

Note that I moved the 'creature type' patching from Actors.Stats into a
new tag, Creatures.Type, since that's not a stat at all. Also added the
tag to Oblivion. This serves as a compromise between not patching
creature types at all (as we currently do for Oblivion) and patching
them as part of Actors.Stats (as valda's version did).