Add MustBeActiveIfImported to all games
...except FO4 and SSE, since it's pointless there. Also added Deactivate
and Filter to FO4, though I doubt they'll be very useful there.
No reason not to have this. It's a meta tag, so if people want to use
it, they should be able to.

Since this is basically the end of this branch, I sorted the patchers
and allTags variables.
Also finally fixed all the 'in the following records'. Those are record
*types*, not records.

Finally, I dropped ScriptContents from Oblivion - the patcher is long
gone, we don't even document the tag anymore, let's stop recognizing it
altogether.