Drop bundled chardet RRR
This is version 1.0.1, which is from 2008. Slightly outdated, let's
upgrade to 3.0.4, which supports both py3 and py2.

I threw every patcher we have at it, but it didn't change the checksum
of the resulting BP (after adjusting the description to account for the
date change, obviously), so this doesn't seem to change behavior.

Under # 460 <--- RRR