Some work towards wxPython 3.0.2.0
Ellipsize overly long messages in progress dialogs

Needed in wx3, otherwise the progress dialog will keep resizing itself,
which is *very* ugly to look at. A pretty hacky solution, since this
won't hide the resizing for displays smaller than 1080p.

Use pngcrush to avoid iCCP warning

This implements it for wizards, followup commit does the same for
FOMODs.
Also added a licenses.txt file to the Mopy/bash/compiled folder to
satisfy any requirements for attribution.

Add pngcrush executable

This is pngcrush 1.8.11 32 bit from
https://sourceforge.net/projects/pmt/files/pngcrush-executables

It's the last version that a windows executable was uploaded for, 1.8.12
and 1.8.13 do not have official executables so we'd have to build our
own.

TODO: Once we upgrade everything to 64 bit (CBash and LOOT are
stopping us BTW), we should swap this for the 64 bit version as
well, I guess.

First implementation for wizards

- Only wizards
- Only for archives (projects should be fixed by the author)
- Runs on all PNGs (seems to be *really* slow)
- Fixes the issue, however

Stuck it into archives.py right now, but that's not the right place IMO.

Skip PNGs that don't have incorrect profiles

Speeds it up *significantly*, making it pretty much instant for working
PNGs.

Delay progress dialog when unpacking to project

Under wx3, it would shove itself over any warning messages / rename
dialogs we showed. This wasn't really a *problem*, since the user could
just move the progress dialog out of the way, but it was annoying. This
is a fairly painless fix that just asks the questions first and then
creates the progress dialog.

Under # 15 <--- RRR
Closes # 458 <--- RRR

Co-authored-by: MrD <the.ubik@gmail.com>